refactor: flatten project file structure
