+++ src/components/DevName.js
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import { OPENSEA_DIRECT_LINK_PREFIX } from '../../../utils/DeveloperDaoConstants';
+import { OPENSEA_DIRECT_LINK_PREFIX } from '../utils/DeveloperDaoConstants';
 import { Button, VStack } from '@chakra-ui/react';
-import IconOpenSea from '../../Icons/opensea';
+import IconOpenSea from './Icons/opensea';
 
 function DevName({ nft, developerId }) {
   if (nft.owner) {
